Move v8 ResizeGroup changes from react-next to react
