+++ packages/react-internal/src/components/ResizeGroup/ResizeGroup.tsx
@@ -1,3 +1,5 @@
+import * as React from 'react';
 import { ResizeGroupBase } from './ResizeGroup.base';
+import { IResizeGroupProps } from './ResizeGroup.types';
 
-export const ResizeGroup = ResizeGroupBase;
+export const ResizeGroup: React.FunctionComponent<IResizeGroupProps> = ResizeGroupBase;
