--- packages/react-next/src/components/ResizeGroup/ResizeGroup.tsx
@@ -1,3 +1,0 @@
-import { ResizeGroupBase } from './ResizeGroup.base';
-
-export const ResizeGroup = ResizeGroupBase;
