--- packages/react-next/src/components/ResizeGroup/index.ts
@@ -1,3 +1,0 @@
-export * from './ResizeGroup';
-export * from './ResizeGroup.base';
-export * from './ResizeGroup.types';
